Split audio output stream modules
